Enable CI with Github actions (#2)
Enable CI with Github actions

I had some issues with sourcing the virtual environment setup script in the CI environment. It took some iteration, but I finally got it working. The main change to accomplish that was to add an override flag to the error check in `set-up-python-venv.sh`. The way that script is executed when sourced in the workflow YAML file causes it to fail that check. I'm still not 100% certain the script is actually being sourced, however.

At one point as I was working through this, I had thought that maybe it would be better to not use the virtual environment in the CI build, but changed my mind. I think it's best if we keep our local dev, CI, and production environments as clean, isolated, and consistent as possible. Using the virtual environment in all places will help accomplish that.