Fix zsh portability issue in venv script
The venv setup script has to be sourced to work correctly. To check this
and emit an error if run directly, it has to rely upon indirect signals.
We discovered today, when the script is sourced in a pure zsh
environment, the logic falsely derives it is not being sourced.

This commit introduces more robust analysis of the execution environment
to determine how the script was being called. It also adds option
switches for forcing it to run regardless of what it infers as well as
to emit debug information to aid in the case of further issues.