Fix deadline handling in commission test (#41)
`test_commission_art_piece` was `mock`ing the `send_deadline_reached`
callback setup and then `await`ing on the result before asserting that
`set` is called the expected number of times. With the last commit, the
`call_later` call was changed to a different overload, one that does not
call `create_task`. This, in turn, results in one fewer `args` items and
the `deadline_timer` member not being assigned by `patch` as was
previously expected. This initially caused an `IndexError` and then
`await` failed with the following:
`TypeError: object NoneType can't be used in 'await' expression`

This commit removes the items that no longer match the current
expectations, given the change in execution flow.