Fail build on invalid source format (#43)
* Fail build on invalid source format

Source that doesn't comply with our code style is consistently getting
checked in. This commit changes the build verification to fail the build
in that case.

* Run `make ruff-format` on source

I had purposefully skipped this to verify that the build would, in fact,
fail if source was not correctly formatted. It does fail as intended.