Removed duplicate ledger.py file.
There was a duplicate ledger.py file in the ledger package. It makes
more sense to keep the ledger.py file under the peer folder, because
a peer maintains ledgers for their different owned artworks. The
ledger.py file under the ledger folder was more updated, so I copied
that code onto the ledger.py file under the peer folder and removed
the ledger.py file under the ledger folder.